+++ source/extensions/filters/http/ratelimit/config.cc
@@ -10,7 +10,6 @@
 #include "common/protobuf/utility.h"
 
 #include "extensions/filters/common/ratelimit/ratelimit_impl.h"
-#include "extensions/filters/common/ratelimit/ratelimit_registration.h"
 #include "extensions/filters/http/ratelimit/ratelimit.h"
 
 namespace Envoy {
@@ -31,9 +30,8 @@ Http::FilterFactoryCb RateLimitFilterConfig::createFilterFactoryFromProtoTyped(
   return [proto_config, &context, timeout,
           filter_config](Http::FilterChainFactoryCallbacks& callbacks) -> void {
     callbacks.addStreamFilter(std::make_shared<Filter>(
-        filter_config,
-        Filters::Common::RateLimit::rateLimitClient(
-            context, proto_config.rate_limit_service().grpc_service(), timeout)));
+        filter_config, Filters::Common::RateLimit::rateLimitClient(
+                           context, proto_config.rate_limit_service().grpc_service(), timeout)));
   };
 }